+++ sample_caller.py
@@ -16,6 +16,7 @@ import requests
 
 from juriscraper.lib.importer import build_module_list, site_yielder
 from juriscraper.lib.log_tools import make_default_logger
+from juriscraper.lib.ocr_utils import needs_ocr
 from juriscraper.lib.string_utils import trunc
 
 logger = make_default_logger()
@@ -86,8 +87,20 @@ def log_dict(dic: dict) -> None:
         logger.debug('    %s: "%s"', k, v)
 
 
+def extract_content_from_doctor(url, files, ocr_available=False):
+    params = {"ocr_available": ocr_available}
+    response = requests.post(url, files=files, timeout=120, params=params)
+    response.raise_for_status()
+    return response.json()["content"], response.json()["page_count"]
+
+
 def extract_doc_content(
-    data, extract_from_text: bool, site, doctor_host: str, filename: str
+    data,
+    extract_from_text: bool,
+    site,
+    doctor_host: str,
+    filename: str,
+    ocr_available: bool = False,
 ):
     """Extracts document's content using a local doctor host
 
@@ -101,6 +114,7 @@ def extract_doc_content(
     :param doctor_host: local doctor instance host. calls will fail if
         the doctor host is not valid
     :param filename: Name for saving extracted content into a file in tmp
+    :param ocr_available: if True, it will tell doctor that OCR is available
 
     :return: a tuple with:
         the extracted content
@@ -119,9 +133,15 @@ def extract_doc_content(
 
     files = {"file": (f"something.{extension}", data)}
     url = MICROSERVICE_URLS["document-extract"].format(doctor_host)
-    extraction__response = requests.post(url, files=files, timeout=120)
-    extraction__response.raise_for_status()
-    extracted_content = extraction__response.json()["content"]
+    extracted_content, page_count = extract_content_from_doctor(
+        url, files, ocr_available=False
+    )
+
+    if ocr_available and needs_ocr(extracted_content, page_count):
+        logger.info("OCR is needed for this document. Using OCR doctor.")
+        extracted_content, page_count = extract_content_from_doctor(
+            url, files, ocr_available=True
+        )
 
     # The extracted content is embedded for display in Courtlistener.
     # We save it into /tmp/ to have an idea how it would look. You can
@@ -246,6 +266,7 @@ def scrape_court(
     doctor_host="",
     test_hashes: bool = False,
     limit: int = 1000,
+    ocr_available: bool = False,
 ):
     """Calls the requested court(s), gets its binary content, and
     extracts the content if possible. See --extract-content option
@@ -298,7 +319,7 @@ def scrape_court(
         filename = item["case_names"].lower().replace(" ", "_")[:40]
 
         data, metadata_from_text = extract_doc_content(
-            data, extract_content, site, doctor_host, filename
+            data, extract_content, site, doctor_host, filename, ocr_available
         )
         logger.log(
             5, "\nShowing extracted document data (500 chars):\n%s", data[:500]
@@ -486,6 +507,12 @@ def main():
         default=1000,
         help="How many items to scrape per `scrape_court` call",
     )
+    parser.add_option(
+        "--ocr-available",
+        action="store_true",
+        default=False,
+        help="If set it will tell doctor that OCR is available. ",
+    )
 
     (options, args) = parser.parse_args()
 
@@ -501,6 +528,7 @@ def main():
     save_responses = options.save_responses
     test_hashes = options.test_hashes
     limit_per_scrape = options.limit_per_scrape
+    ocr_available = options.ocr_available
 
     if test_hashes:
         binaries = True
@@ -572,6 +600,7 @@ def main():
                     doctor_host,
                     test_hashes,
                     limit_per_scrape,
+                    ocr_available,
                 )
 
     logger.debug("The scraper has stopped.")